Various wxJS bug fixes and minor feature additions
git-svn-id: http://svn.wildfiregames.com/public/ps/trunk@5324 3db68df2-c116-0410-a063-a993310a9797

+++ source/tools/atlas/wxJS/common/apiwrap.h
@@ -750,7 +750,7 @@ JSClass wxjs::ApiWrapper<T_Port, T_Priv>::wxjs_class =
     }
 
 // Defines a constant with a prefix
-#define WXJS_CONSTANT(prefix, name)	{ prefix##name, #name, WXJS_READONLY, { 0 } },
+#define WXJS_CONSTANT(prefix, name)	{ (int)prefix##name, #name, WXJS_READONLY, { 0 } },
 // Defines a constant
 #define WXJS_SIMPLE_CONSTANT(name)  { name, #name, WXJS_READONLY, { 0 } },
